feat(lists): shared query bar with multi-select filters for spots & tags
Extract the search/filter/sort UI into a shared ListQueryBar backed by a configurable PickerOverlay. List filters are now multi-select and combine as an intersection (e.g. Map Extent + Recent Views). Renamed the spot/tag "filters" modules to "query" to match; tag lists always group by type.
