+++ src/modules/spots/spots.constants.js
@@ -14,14 +14,12 @@ export const SORT_ORDER = {
   RECENTLY_VIEWED: 'Recently Viewed',
 };
 
-export const SORTED_VIEWS = {
-  CHRONOLOGICAL: 'CHRONOLOGICAL',
+export const FILTERS = {
   MAP_EXTENT: 'MAP_EXTENT',
   RECENT_VIEWS: 'RECENT_VIEWS',
 };
 
-export const SORTED_VIEW_LABELS = {
-  [SORTED_VIEWS.CHRONOLOGICAL]: 'In Active Datasets',
-  [SORTED_VIEWS.MAP_EXTENT]: 'In Map Extent',
-  [SORTED_VIEWS.RECENT_VIEWS]: 'In Recent Views',
+export const FILTER_LABELS = {
+  [FILTERS.MAP_EXTENT]: 'Map Extent',
+  [FILTERS.RECENT_VIEWS]: 'Recent Views',
 };
