+++ src/modules/tags/Tags.js
@@ -4,8 +4,9 @@ import {View} from 'react-native';
 import {useDispatch, useSelector} from 'react-redux';
 
 import BackupTagsModal from './BackupTagsModal';
-import TagFilters from './filters/TagFilters';
 import LoadTagsModal from './LoadTagsModal';
+import TagQuery from './query/TagQuery';
+import {TAG_FILTERS} from './query/tagQuery.constants';
 import {TAG_TYPES} from './tags.constants';
 import TagsOverflowMenuModal from './TagsOverflowMenuModal';
 import {isEmpty} from '../../shared/helpers';
@@ -47,8 +48,9 @@ const Tags = ({
   const label = page.label;
   const addButtonTitle = isGeologicUnits ? 'Create New Geologic Unit' : 'Create New Tag';
 
-  // Each page (Tags/Geologic Units) keeps its own filter in Redux so selecting one doesn't affect the other.
-  const selectedIndex = (listFilters && listFilters[pageKey]) || 0;
+  // Each page (Tags/Geologic Units) keeps its own filters in Redux so selecting one doesn't affect the other.
+  const pageFilter = listFilters?.[pageKey];
+  const activeFilters = Array.isArray(pageFilter) ? pageFilter : [];
 
   /* Derived State */
 
@@ -61,11 +63,11 @@ const Tags = ({
 
   // Let the map know a map-extent list is being viewed so it auto-recomputes the extent on move.
   useEffect(() => {
-    if (selectedIndex === 1) {
+    if (activeFilters.includes(TAG_FILTERS.MAP_EXTENT)) {
       dispatch(setIsMapExtentFilterActive(true));
       return () => dispatch(setIsMapExtentFilterActive(false));
     }
-  }, [selectedIndex]);
+  }, [pageFilter]);
 
   /* Event Handlers */
 
@@ -81,7 +83,7 @@ const Tags = ({
 
   const closeDetailModal = () => setIsDetailModalVisible(false);
 
-  const setSelectedIndex = index => dispatch(setListFilters({page: pageKey, value: index}));
+  const setActiveFilters = filters => dispatch(setListFilters({page: pageKey, value: filters}));
 
   /* View */
 
@@ -90,16 +92,16 @@ const Tags = ({
       <AddButton onPress={addTag} title={addButtonTitle}/>
       {!isEmpty(tags) && (
         <>
-          <TagFilters
+          <TagQuery
+            activeFilters={activeFilters}
             isGeologicUnits={isGeologicUnits}
-            selectedIndex={selectedIndex}
-            setSelectedIndex={setSelectedIndex}
+            setActiveFilters={setActiveFilters}
             setTagsSorted={setTagsSorted}
             tags={baseTags}
           />
         </>
       )}
-      <TagsList selectedIndex={selectedIndex} tagsSorted={tagsSorted} type={pageKey}/>
+      <TagsList activeFilters={activeFilters} tagsSorted={tagsSorted} type={pageKey}/>
 
       {/* Menus and Modals */}
       <TagsOverflowMenuModal
