+++ src/modules/tags/TagsList.js
@@ -1,11 +1,12 @@
 import React from 'react';
-import {FlatList, SectionList, Text, View} from 'react-native';
+import {SectionList, Text, View} from 'react-native';
 
 import {ListItem} from '@rn-vui/base';
 import {useDispatch, useSelector} from 'react-redux';
 
 import {useTags} from '.';
 import TagColorIcon from './color/TagColorIcon';
+import {TAG_FILTERS} from './query/tagQuery.constants';
 import {TAG_SECTIONS} from './tags.constants';
 import {getTagTitle} from './tags.helpers';
 import commonStyles from '../../shared/common.styles';
@@ -19,7 +20,7 @@ import {PRIMARY_PAGES} from '../page/page.constants';
 import {PAGE_KEYS} from '../page/pageKeys.constants';
 import {setSelectedTag} from '../project/projects.slice';
 
-const TagsList = ({selectedIndex, tagsSorted, type}) => {
+const TagsList = ({activeFilters = [], tagsSorted, type}) => {
   console.log('Rendering TagsList...');
 
   const dispatch = useDispatch();
@@ -69,36 +70,25 @@ const TagsList = ({selectedIndex, tagsSorted, type}) => {
     );
   };
 
-  const renderTagsListByMapExtent = () => {
-    const tagsInMapExtent = (tagsSorted || []).filter(
-      tag => tag.spots && !isEmpty(tag.spots.find(spotId => spotsInMapExtentIds?.includes(spotId))),
-    );
-    console.log('tagsInMapExtent', tagsInMapExtent);
-
-    return (
-      <FlatList
-        ItemSeparatorComponent={FlatListItemSeparator}
-        ListEmptyComponent={<ListEmptyText text={`No Spots with ${label.toLowerCase()} in current map extent`}/>}
-        data={tagsInMapExtent}
-        keyExtractor={item => item.id.toString()}
-        renderItem={({item}) => renderTag(item)}
-      />
-    );
-  };
-
-  const renderTagsListByRecentlyUsed = () => {
-    return <ListEmptyText text={'This has not been implemented yet'}/>;
-  };
-
-  const renderTagsListByType = () => {
+  // Always grouped by type; active filters (e.g. Map Extent) narrow the tags as an intersection.
+  const renderTagsList = () => {
+    const hasActiveFilters = !isEmpty(activeFilters);
+    const inMapExtent = activeFilters.includes(TAG_FILTERS.MAP_EXTENT);
     const dataSectioned = Object.values(SECTIONS).reduce((acc, {title, key}) => {
-      const data = (tagsSorted || []).filter(d => d.type === key);
-      return [...acc, {title: title, data: data}];
+      let data = (tagsSorted || []).filter(d => d.type === key);
+      if (inMapExtent) {
+        data = data.filter(
+          tag => tag.spots && !isEmpty(tag.spots.find(spotId => spotsInMapExtentIds?.includes(spotId))),
+        );
+      }
+      // With filters active, hide empty type sections; unfiltered, keep them (they show a "No <type>" footer).
+      return hasActiveFilters && isEmpty(data) ? acc : [...acc, {title: title, data: data}];
     }, []);
 
     return (
       <SectionList
         ItemSeparatorComponent={FlatListItemSeparator}
+        ListEmptyComponent={<ListEmptyText text={`No ${label.toLowerCase()} match the current filters`}/>}
         keyExtractor={(item, index) => item + index}
         renderItem={({item}) => renderTag(item)}
         renderSectionFooter={({section: {data, title}}) => {
@@ -119,15 +109,13 @@ const TagsList = ({selectedIndex, tagsSorted, type}) => {
   else {
     return (
       <>
-        <View style={{alignItems: 'flex-end', paddingHorizontal: 10}}>
+        <View style={{alignItems: 'flex-end', paddingHorizontal: 10, paddingTop: 10}}>
           <Text style={commonStyles.standardDescriptionText}>
             {pageKey === PAGE_KEYS.GEOLOGIC_UNITS ? '# Tagged Spots | Samples' : '# Tagged Spots | Samples | Features'}
           </Text>
         </View>
         <View style={{flex: 1}}>
-          {selectedIndex === 0 && renderTagsListByType()}
-          {selectedIndex === 1 && renderTagsListByMapExtent()}
-          {selectedIndex === 2 && renderTagsListByRecentlyUsed()}
+          {renderTagsList()}
         </View>
       </>
     );
